added convenience function for unwrapping config to replace commonly used code
