+++ transformer_lens/components/abstract_attention.py
@@ -43,9 +43,7 @@ class AbstractAttention(ABC, nn.Module):
             layer_id (int, optional): The index of the current layer. Used by the Mistral models (labelled here as stanford-gpt2) to scale down attention scores pre softmax for numerical stability reasons by 1/(layer_id+1). Defaults to None.
         """
         super().__init__()
-        if isinstance(cfg, Dict):
-            cfg = HookedTransformerConfig.from_dict(cfg)
-        self.cfg = cfg
+        self.cfg = HookedTransformerConfig.unwrap(cfg)
 
         if self.cfg.load_in_4bit:
             nq = int((cfg.d_model * cfg.d_model) / 2)
