+++ transformer_lens/components/attention.py
@@ -33,9 +33,7 @@ class Attention(AbstractAttention):
             layer_id (int, optional): The index of the current layer. Used by the Mistal models (labelled here as stanford-gpt2) to scale down attention scores pre softmax for numerical stability reasons by 1/(layer_id+1). Defaults to None.
         """
         super().__init__(cfg, attn_type, layer_id)
-        if isinstance(cfg, Dict):
-            cfg = HookedTransformerConfig.from_dict(cfg)
-        self.cfg = cfg
+        self.cfg = HookedTransformerConfig.unwrap(cfg)
 
         if cfg.load_in_4bit:
             # 4-bit quantization convention
